move TagEventJS back to its original place
